movie_app/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
bcrypt==4.0.1
Werkzeug==3.1.3